+++ app/assets/javascripts/components/admin/animes/detail/seasons/_admin_anime_season_new_field.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import { origin } from './../../../../../origin.js'
-import AdminNewButtonField from './../../../_admin_new_button_field'
+import AdminNewButtonField from './../../../common/_admin_new_button_field'
 import AdminAnimeSeasonForm from './_admin_anime_season_form'
 
 export default class AdminAnimeSeasonNewField extends Component {
