+++ app/assets/javascripts/components/admin/animes/list/_admin_anime_new_form.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import { origin } from './../../../../origin.js'
 import MessageBox from './../../../common/_message_box'
-import AdminNewButtonField from './../../_admin_new_button_field'
+import AdminNewButtonField from './../../common/_admin_new_button_field'
 
 export default class AdminAnimeNewForm extends Component {
   constructor(props) {
