+++ app/assets/javascripts/components/admin/common/_admin_advertisement_form.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
-import MessageBox from './../../../../common/_message_box'
+import MessageBox from './../../common/_message_box'
 
 export default class AdminAdvertisementForm extends Component {
   constructor(props) {
